tg-hub-src/requirements.txt
telethon==1.36.0
python-telegram-bot==21.5
anthropic>=0.96.0
apscheduler==3.10.4
pyyaml==6.0.2
pytz==2024.1
aiofiles==23.2.1
flask==3.0.3
pyotp==2.9.0
qrcode[pil]==7.4.2
